bot/requirements.txt
aiogram>=3.15,<4
fastapi>=0.115,<1
uvicorn>=0.32,<1
sqlalchemy[asyncio]>=2.0,<3
alembic>=1.13,<2
asyncpg>=0.30,<1
pydantic-settings>=2.0,<3
grpcio>=1.60,<2
protobuf>=4.25,<6
pyyaml>=6.0,<7
